1. Create `connected` components folder 2. Separate SongList UI from connection 3. Update component locations
