+++ src/connected/ArtistPage.js
