+++ src/connected/ArtistSongList.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import * as selectors from '../redux/selectors';
-import { fetchArtist } from '../redux/actions';
-import SongList from './SongList';
 import { connect } from 'react-redux';
+import { addSongs, setCurrentSongId } from '../redux/actions';
+import * as selectors from '../redux/selectors';
+import SongList from '../components/SongList';
 
 const ArtistSongList = (props) => {
 	return <SongList {...props} />;
@@ -12,4 +12,4 @@ const mapState = (state, ownProps) => ({
 	songs: selectors.getArtistsSongs(state, ownProps.artistId)
 });
 
-export default connect(mapState, { fetchArtist })(ArtistSongList);
+export default connect(mapState, { addSongs, setCurrentSongId })(ArtistSongList);
